Create dis.play.php
The full attention check is 3 parts (or web pages) in Collector: (1) instructions/what you would like to be checked, (2) a question that tests whether participants are paying attention (3) either a new web page informing the participant that they failed the check and thus cannot continue the study (if they answered the question incorrectly or did not answer before hitting submit) OR a page that simply says “Processing….” and automatically moves on to the next part of the experiment (the next row in the procedure file) if they answer the question correctly.

Some helpful links:
To see a toy example of the attention check: www.cogfog.com/SP/test01 (use condition 15)
Where participants are redirected when they fail an attention check: www.cogfoglab.wixsite.com/website

To use the attention check trial type (AttnCheck):
Directions/instructions (whatever information you would like to be checked) should be in their own separate row in the procedure file. 
The following procedure row should be a multiple choice question that asks the participant about the previous information. This is the attention check question, but its trial type should be multipleChoiceText. It can be timed or user paced.
In the procedure row immediately following the question, the trial type is AttnCheck. Fill out the “text” column in this row with the correct answer to the attention check question. AttnCheck compares the text inputted here with the answer selected in the previous part. Thus, spelling, spacing, and format should be exactly the same as what the answer choice is in the question. Make sure to change timing (Max Time) on this procedure row to 2 seconds. 
Note: the timing in this row determines how long the participants who answered the attention check question correctly will see the “Processing…” screen before they are automatically taken to the next question/continue in the study.
Make sure shuffle is off for these 3 rows of the procedure.